Fix rpm 4.15+ compatibility
Before rpm 4.15, rpm's python3 bindings would return string data as
bytes.  This script worked around that behavior by proactively decoding
them, which was a noop on python2, retaining compatibility.  In rpm 4.15
this behavior was corrected so that string data is returned as utf8
strings, which had a side effect of breaking our workaround on python3.
The change was also backported to RHEL as rpm-4.14.2-25.el8, available
since RHEL 8.1.  The last Fedora release with the old behavior was
Fedora 30, meaning all OS's with the bytes behavior are EOL.  We can
just use the strings as is, which works on all current Fedora and RHEL
releases.

https://github.com/jds2001/keychecker/commit/da993741f78d873f6619384a2c857ec6710aa2fb
https://github.com/rpm-software-management/rpm/commit/84920f898315d09a57a3f1067433eaeb7de5e830
https://access.redhat.com/errata/RHBA-2019:3584